requirements.txt
websockets==12.0
gtts==2.4.0
pygame==2.5.2
obs-websocket-py==1.0
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
groq==0.4.1
asyncio
